Removed redundant lines, fixed checking entirely
All instances of total_count have been removed or replaced by a check on the length of the list of elementals. This list is appended to anyways and as such, the total_count variable was redundant. The program is now 5 lines shorter for it.

The in-submission checks for repetition, a minimum of one base elemental, and a maximum of one celestial elemental have been implemented fully. They've been overhauled into using while loops, meaning that repeated input by a user cannot bypass them as it could the previous if statements.

Aether was added as a delta elemental as well as a celestial elemental.

A print statement was established at the beginning of the pick3 function going over valid inputs and the required formatting to make the function work.